+++ modules/azure/vault/vault.tf
@@ -1,3 +1,7 @@
+locals {
+  role_definition_id = uuid()
+}
+
 resource "azurerm_managed_disk" "vault_data" {
   name                 = "${var.cluster_name}-vault-data-disk"
   location             = var.location
@@ -100,6 +104,7 @@ resource "azurerm_virtual_machine" "vault" {
 }
 
 resource "azurerm_role_definition" "vault_access_role" {
+  role_definition_id = local.role_definition_id
   name        = "${var.cluster_name}-vault-access"
   scope       = "/subscriptions/${var.subscription_id}/resourceGroups/${var.cluster_name}"
   description = "Custom role used to provide vault access to VMs/VMSSs"
@@ -107,15 +112,11 @@ resource "azurerm_role_definition" "vault_access_role" {
   permissions {
     actions = ["Microsoft.Compute/virtualMachineScaleSets/read", "Microsoft.Compute/virtualMachines/read"]
   }
-}
 
-# there is delay between role is visible for assignment
-resource "null_resource" "delay_role_assignment_creation" {
-  provisioner "local-exec" {
-    command = "sleep 30"
-  }
-  triggers = {
-    "before" = azurerm_role_definition.vault_access_role.id
+  lifecycle {
+    ignore_changes = [
+      role_definition_id,
+    ]
   }
 }
 
@@ -125,7 +126,6 @@ resource "azurerm_role_assignment" "vault_access_role_assignment" {
   principal_id         = azurerm_virtual_machine.vault.identity[0].principal_id
 
   depends_on = [
-    azurerm_role_assignment.vault_access_role_assignment,
-    null_resource.delay_role_assignment_creation,
+    azurerm_role_definition.vault_access_role,
   ]
 }